some more PROGMEM related fixes

+++ ArduPar3Collection.cpp
@@ -22,13 +22,13 @@ size_t ArduPar3Collection::registerInstance(AbstractArduPar3 *instance)
 /// write information about all parameter instances to a stream
 void ArduPar3Collection::dumpParameterInfos(Stream *outStream)
 {
-  for (int i = 0; i < numInstancesRegistered; i++)
+  for (size_t i = 0; i < numInstancesRegistered; i++)
   {
     knownInstances[i]->dumpParameterInfo(outStream);
   }
 };
 /// this function automatically distributes incoming data from a stream to the parameter setting instances
-void ArduPar3Collection::updateParametersFromStream(Stream *inStream, int timeout)
+void ArduPar3Collection::updateParametersFromStream(Stream *inStream, unsigned long timeout)
 {
   if (!inStream->available())
     return; // if there is no data available, dont even start timeouts
@@ -36,7 +36,7 @@ void ArduPar3Collection::updateParametersFromStream(Stream *inStream, int timeou
   char buf[PAR_SETTINGS_BUFFER_SIZE];
   int bufPos = 0;
   int inByte;
-  long lastCharMillis = millis();
+  unsigned long lastCharMillis = millis();
   // even if the buffer is full, we keep on reading to flush the remaining data.
   bool receivedNewLine = false;
   while ((millis() - lastCharMillis) < timeout && !receivedNewLine)
@@ -60,7 +60,7 @@ void ArduPar3Collection::updateParametersFromStream(Stream *inStream, int timeou
   // let all the parameter setting instances have a look at the data
   if (bufPos > 0)
   {
-    for (int i = 0; i < numInstancesRegistered; i++)
+    for (size_t i = 0; i < numInstancesRegistered; i++)
     {
       knownInstances[i]->parseCommand(buf);
     }
@@ -68,13 +68,13 @@ void ArduPar3Collection::updateParametersFromStream(Stream *inStream, int timeou
 };
 
    void ArduPar3Collection::loadAll(){
-    for (int i = 0; i < numInstancesRegistered; i++)
+    for (size_t i = 0; i < numInstancesRegistered; i++)
     {
       knownInstances[i]->load();
     }
    }
    void ArduPar3Collection::saveAll(){
-    for (int i = 0; i < numInstancesRegistered; i++)
+    for (size_t i = 0; i < numInstancesRegistered; i++)
     {
       knownInstances[i]->save();
     }
